Fixed the app to work with Flutter 3.13.x.
I was having trouble with this sample app not working with build errors in Flutter 3.13.x., so I created a new one and made the same app.

We have also added a process to hide the debug burner that appears in the application.

Please check your programs for problems and merge them.
We believe that this fix will help Flutter beginners learn Flutter smoothly.